+++ js/tests/test_timepicker.js
@@ -1,4 +1,6 @@
 import {TimepickerWidget} from '../src/timepicker.js';
+import {register_timepicker_array_subscribers} from '../src/timepicker.js';
+
 
 QUnit.module('TimepickerWidget', hooks => {
     let container = $('<div id="container" />');
@@ -6,6 +8,9 @@ QUnit.module('TimepickerWidget', hooks => {
     let data_clock;
     let data_locale;
     let picker;
+    let _array_subscribers = {
+        on_add: []
+    };
 
     hooks.before(() => {
         $('body').append(container);
@@ -19,6 +24,51 @@ QUnit.module('TimepickerWidget', hooks => {
         // reset container css
         container.css('top', 'unset').css('left', 'unset').css('position', 'unset');
         picker = null;
+    });
+
+    QUnit.test('register_timepicker_array_subscribers', assert => {
+        container.empty();
+
+        // return if window.yafowil === undefined
+        register_timepicker_array_subscribers();
+        assert.deepEqual(_array_subscribers['on_add'], []);
+
+        // patch yafowil_array
+        window.yafowil_array = {
+            on_array_event: function(evt_name, evt_function) {
+                _array_subscribers[evt_name] = evt_function;
+            },
+            inside_template(elem) {
+                return elem.parents('.arraytemplate').length > 0;
+            }
+        };
+        register_timepicker_array_subscribers();
+
+        // create table DOM
+        let table = $('<table />')
+            .append($('<tr />'))
+            .append($('<td />'))
+            .appendTo('body');
+        
+        let el = $(`<input type="text" />`).addClass('timepicker');
+        $('td', table).addClass('arraytemplate');
+        el.appendTo($('td', table));
+
+        // invoke array on_add - returns
+        _array_subscribers['on_add'].apply(null, $('tr', table));
+        picker = el.data('yafowil-timepicker');
+        assert.notOk(picker);
+        $('td', table).removeClass('arraytemplate');
+
+        // invoke array on_add
+        el.attr('id', '');
+        _array_subscribers['on_add'].apply(null, $('tr', table));
+        picker = el.data('yafowil-timepicker');
+        assert.ok(picker);
+
+        table.remove();
+        window.yafowil_array = undefined;
+        _array_subscribers = undefined;
     });
 
     QUnit.test('Initialize() - no data', assert => {
@@ -480,6 +530,21 @@ QUnit.module('TimepickerWidget', hooks => {
         // click inside of element
         picker.dd_elem.trigger('click');
         assert.strictEqual(picker.dd_elem.css('display'), 'block');
+    });
+
+    QUnit.test('change event', assert => {
+        TimepickerWidget.initialize();
+        picker = elem.data('yafowil-timepicker');
+        picker.elem.on('change', () => {
+            assert.step('timepicker_input_change');
+        })
+
+        picker.trigger_elem.trigger('click');
+        picker.minute = '25';
+        picker.hour = '14';
+        picker.set_time();
+        assert.verifySteps(['timepicker_input_change']);
+        picker.elem.off('change');
     });
 });